requirements.txt
aiohttp==3.8.3
async-timeout==4.0.2
attrs==22.2.0
chardet==5.1.0
configparser==5.3.0
discord==2.1.0
discord.py==2.1.0
idna==3.4
multidict==6.0.4
python-dotenv==0.21.0
typing-extensions==4.4.0
yarl==1.8.2
